Pubmed/requirements.txt
streamlit
google-cloud-bigquery
google-cloud-translate
pandas
db-dtypes
pyarrow

